# openapi.yaml
openapi: 3.1.0
info:
  title: Form Engine API
  version: 0.1.0
  description: |
    Dynamic Form Engine — configuration-driven form building and approval workflows.

    All routes use the `/api/v1` prefix (see ADR-0005). This document is the
    hand-written API contract (spec-first, see ADR-0007) from which client/server
    TypeScript types are generated via `openapi-typescript`.

    SSE streams (notification push, work order 07) are intentionally not modelled
    here — OpenAPI 3.x has weak SSE support; they are described separately.
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/v1/instances:
    post:
      tags: [instances]
      summary: Create a draft instance
      description: Starts a new form for a published template. The instance begins in `draft` status.
      operationId: createInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstanceRequest"
      responses:
        "201":
          description: Draft instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "400":
          description: Template not published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/instances/my:
    get:
      tags: [instances]
      summary: List my instances
      description: |
        Drafts and submissions owned by the acting user, newest first. Optional
        comma-separated status filter.
      operationId: listMyInstances
      parameters:
        - name: status
          in: query
          schema: { type: string }
          description: Comma-separated status filter, e.g. `draft,submitted`.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of the user's instances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceListResponse"
  /api/v1/instances/{id}:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    get:
      tags: [instances]
      summary: Get instance detail
      description: |
        Returns the instance plus its approval records and the live template.
        Drafts render against `template.schema`; submitted instances carry a frozen
        `template_snapshot`.
      operationId: getInstance
      responses:
        "200":
          description: Instance found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "404":
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/instances/{id}/values:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    put:
      tags: [instances]
      summary: Autosave field values
      description: |
        Persists field values on a draft instance without bumping the optimistic
        lock version. Rejects non-draft instances.
      operationId: saveInstanceValues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveValuesRequest"
      responses:
        "200":
          description: Values saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormInstance"
        "400":
          description: Instance is not in draft status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not the owner of the instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/instances/{id}/submit:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    post:
      tags: [instances]
      summary: Submit an instance
      description: |
        Validates the field values against the current template schema and
        atomically submits (ADR-0001): the instance update, frozen snapshot, and
        approval-record inserts commit together; approver resolution runs inside
        that transaction. Notifications persist async after commit.
      operationId: submitInstance
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRequest"
      responses:
        "200":
          description: Submitted (status submitted, or approved when no approvers)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "400":
          description: Not a draft/withdrawn instance or template not published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not the owner of the instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Field validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Approver resolution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/instances/{id}/withdraw:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    post:
      tags: [instances]
      summary: Withdraw a submitted instance
      description: |
        Reverts a `submitted`/`in_approval` instance back to draft. Optimistic
        locking on `version` (409 VERSION_CONFLICT); rejected if any approver has
        already acted (400 APPROVAL_NOT_PENDING).
      operationId: withdrawInstance
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
      responses:
        "200":
          description: Withdrawn (back to draft)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "400":
          description: Not a withdrawable status, or an approver already acted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not the owner of the instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Optimistic-lock version conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      description: Standard error shape returned by the unified error handler and 404 catch-all.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code, e.g. NOT_FOUND, VALIDATION_ERROR.
            message:
              type: string
              description: Human-readable message (Chinese in the MVP).
...
    FormInstance:
      type: object
      description: A filled form submission at runtime. Mirrors the form_instances table.
      required: [id, template_id, template_snapshot, field_values, status, current_node_index, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        template_snapshot:
          type: object
          additionalProperties: true
          description: Frozen copy of the template schema at submission time.
        field_values:
          type: object
          additionalProperties: true
          description: User-filled field values (JSONB).
        status:
          type: string
          enum: [draft, submitted, in_approval, approved, rejected, returned, withdrawn]
        current_node_index:
          type: integer
          description: Current approval node index (0-based).
        version:
          type: integer
          description: Optimistic-locking version.
        submitted_by:
          type: [string, "null"]
          format: uuid
        submitted_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TemplateSummary:
      type: object
      description: A reduced template view embedded in instance/draft detail responses.
      required: [id, name, status, schema, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        schema:
          type: object
          additionalProperties: true
        approval_chain:
          type: [object, "null"]
        updated_at:
          type: string
          format: date-time
...
    ApprovalRecordSummary:
      type: object
      description: An approval record with the approver's display name resolved.
      required: [id, node_id, node_order, action]
      properties:
        id:
          type: string
          format: uuid
        node_id:
          type: string
        node_order:
          type: integer
        approver_id:
          type: [string, "null"]
          format: uuid
        approver_name:
          type: [string, "null"]
        action:
          type: string
          enum: [pending, approved, rejected, returned, transferred]
        comment:
          type: [string, "null"]
        acted_at:
          type: [string, "null"]
          format: date-time
    InstanceDetail:
      type: object
      description: A FormInstance enriched with approval progress and the live template.
      allOf:
        - $ref: "#/components/schemas/FormInstance"
        - type: object
          required: [approval_records, template]
          properties:
            approval_records:
              type: array
              items:
                $ref: "#/components/schemas/ApprovalRecordSummary"
            template:
              $ref: "#/components/schemas/TemplateSummary"
    InstanceListItem:
      type: object
      description: A FormInstance with the template display name resolved.
      allOf:
        - $ref: "#/components/schemas/FormInstance"
        - type: object
          properties:
            template_name:
              type: [string, "null"]
    InstanceListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InstanceListItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    CreateInstanceRequest:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
          format: uuid
        field_values:
          type: object
          additionalProperties: true
          description: Optional initial field values.
    SaveValuesRequest:
      type: object
      required: [field_values]
      properties:
        field_values:
          type: object
          additionalProperties: true
    SubmitRequest:
      type: object
      properties:
        field_values:
          type: object
          additionalProperties: true
          description: Final values to submit; defaults to the last autosaved values.
    WithdrawRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          description: Expected optimistic-lock version (required); a mismatch yields 409 VERSION_CONFLICT.
...
  parameters:
...
    InstanceId:
      name: id
      in: path
      required: true
      description: Form instance UUID.
      schema:
        type: string
        format: uuid
...